Path fixes 🔨
